20151026 page 353 back

+++ backup/PictureUtils.java
@@ -5,21 +5,23 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.drawable.BitmapDrawable;
 import android.view.Display;
-
 import android.widget.ImageView;
 
+/**
+ * Created by yalin on 10/22/2015.
+ */
 public class PictureUtils {
     /**
      * Get a BitmapDrawable from a local file that is scaled down
-     * to fit the current Window size.
+     * to fit the current Window size
      */
     @SuppressWarnings("deprecation")
-    public static BitmapDrawable getScaledDrawable(Activity a, String path) {
+    public  static BitmapDrawable getScaledDrawable(Activity a, String path) {
         Display display = a.getWindowManager().getDefaultDisplay();
         float destWidth = display.getWidth();
         float destHeight = display.getHeight();
 
-        // read in the dimensions of the image on disk
+        //Read in the dimensions of the image on disk
         BitmapFactory.Options options = new BitmapFactory.Options();
         options.inJustDecodeBounds = true;
         BitmapFactory.decodeFile(path, options);
@@ -29,10 +31,10 @@ public class PictureUtils {
 
         int inSampleSize = 1;
         if (srcHeight > destHeight || srcWidth > destWidth) {
-            if (srcWidth > srcHeight) {
-                inSampleSize = Math.round((float)srcHeight / (float)destHeight);
-            } else {
-                inSampleSize = Math.round((float)srcWidth / (float)destWidth);
+            if(srcWidth > srcHeight) {
+                inSampleSize = Math.round(srcHeight/destHeight);
+            }else {
+                inSampleSize = Math.round(srcWidth/destWidth);
             }
         }
 
@@ -44,13 +46,11 @@ public class PictureUtils {
     }
 
     public static void cleanImageView(ImageView imageView) {
-        if (!(imageView.getDrawable() instanceof BitmapDrawable)) 
+        if(!(imageView.getDrawable() instanceof BitmapDrawable))
             return;
-
-        // clean up the view's image for the sake of memory
+        //Clean up the view's image for the sake of memory
         BitmapDrawable b = (BitmapDrawable)imageView.getDrawable();
         b.getBitmap().recycle();
         imageView.setImageDrawable(null);
     }
 }
-
